refactor: extract usePluginsSkillsState hook from AppContext
Move ~40 plugin/skill/catalog/store useState calls + ~15 callbacks into
usePluginsSkillsState. Covers plugin toggle/config, skill CRUD/review,
catalog browse/install, store browse/install, marketplace search.

AppContext.tsx: 7,788 → 7,439 lines (-349)
usePluginsSkillsState.ts: 616 lines (new)
Cumulative: 8,366 → 7,439 (-927 lines, 6 hooks extracted)
